Initial commit in vim repo
This is the initial commit from the previous vim repo hosted at
https://git.bracken.jp/vim.git, formerly a separate repo from my
dotfiles repo hosted at https://git.bracken.jp/dotfiles.git.

Since I always use them together, this merges the two into a single
repo. In particular, with neovim using an XDG_CONFIG_DIR-based config,
it makes sense to merge the two repos so everything is under ~/.config.

This is the start of a replay of the vim repo's history into the
dotfiles repo.