Ensure deterministic English, Japanese terms
Previously, we took the last three terms of `list(set(japanese))` which
is non-deterministic. We now take the last three terms of the sorted set
of the Japanese vocabulary words to ensure a deterministic result. A
similar fix applies to the English terms.

This change allows us to repeatedly call `kotd` as frequently as desired
with no change to the output other than once per `word_time`.