Set fonts.conf alias bindings to strong
This forces priority of configuration rules.

From the Font Matching section of the fontconfig user manual:

    Family names are split into two bindings; strong and weak. Strong
    family names are given greater precedence in the match than lang
    elements while weak family names are given lower precedence than
    lang elements. This permits the document language to drive font
    selection when any document specified font is unavailable.

I'm not convinced this is actually the "correct" solution, but it does
appear to work. It looks like FreeBSD already appends an emoji family to
all other families in /usr/local/etc/fonts/conf.d/45-generic.conf, but
that still doesn't override the default DejaVu bindings.