Eliminate hardcoding of dmenu font
This was added in 4290347 in response to font config changes made in
85e35e0, where the preferred font for serif, sans-serif, and monospace
were set to the Twitter Color Emoji font in an attempt to get emoji
working. That patch alone got most emoji working, but the DejaVu fonts
include a few black-and white glyphs for some basic emoji, which were
taking precedence, since DejaVu fonts are set as the highest strong
bindings to the above-listed aliases.

In 1e1e922, I made the preferred font bindings strong for the Twitter
Color Emoji font, which overrode DejaVu, but made Twitter Color Emoji
the default font for everything on the system. This actually *mostly*
worked because any glyph that wasn't an emoji would just fall back to
the next-highest match, but appears to have mangled font spacing in
monospace fonts.

I fixed all of this in 3a4fb62 by making Noto fonts (which don't include
any emoji codepoints, but do have otherwise broad coverage) to be the
highest-precendence fonts for each alias and Twitter Color Emoji to be
the second-best match. This appears to have resolved the issue
altogether.