Set TERM=xterm-256color in Alacritty
Alacritty uses terminal type 'alacritty' if it exists in the host's
terminfo database, otherwise it falls back to xterm-256color. This is
problematic when terminal type 'alacritty' exists on the local machine
but not a remote host that you're ssh-ing into, in which case, the
connection needs to be made as:

    TERM=xterm-256color ssh user@remote.net

Since xterm-256color covers all my needs, as far as I know, and since
alacritty isn't common enough to be included in terminfo files for most
operating systems, just default to xterm-256colors for now.