Disable ligatures in Noto Sans Mono
Noto Sans Mono includes ligatures for 'fi' etc. which look out-of-place
in a monospace font, but also reduce the number of glyphs in a line by
one for each such ligature. This disables ligatures in all apps that use
fontconfig.

Unfortunately, firefox ignores the `fontfeatures` tag, so the only
option there is to switch fonts.