mbsync: Set CopyArrivalDate yes
This prevents mbsync from setting the email date to the last update date
if you move the message from one IMAP folder to another. I'm not sure if
this is a gmail bug or an mbsync 'feature', but either way, setting
`CopyArrivalDate yes` on each channel apparently fixes the issue.

From the manual:

> Selects whether their arrival time should be propagated together with
> the messages. Enabling this makes sense in order to keep the time
> stamp based message sorting intact. Note that IMAP does not guarantee
> that the time stamp (termed internal date) is actually the arrival
> time, but it is usually close enough. (Default: no)

References:
* https://isync.sourceforge.io/mbsync.html
* https://vxlabs.com/2021/03/21/mbsync-copyarrivaldate-yes/