Make en_CA.UTF-8 locale default other than money
Both FreeBSD and Linux understand that Canadian spelling is just UK
spelling with -ize endings (which are what the OED prefers for that
matter).

Unfortunately macOS/iOS assume that Canadian English is US English.
Google products, for the most part, ignore the existence of Canadian
English, but do recognize Canadian French, leading to lovely behaviours
like YouTube defaulting to Français (Canada) when viewed with a browser
set to English (Canada). *shrug*

This sets the values back to their proper values since, for the most
part, I don't use macOS if I can avoid it.