lynx: default to UTF-8, fix LYNX_CFG env var
The HTTP 1.1 RFC [1] section 3.7.1 says that the default charset is
assumed to be ISO-8859-1:

   When no explicit charset parameter is provided by the sender, media
   subtypes of the "text" type are defined to have a default charset
   value of "ISO-8859-1" when received via HTTP.

On the modern web and gopher sites though, UTF-8 is a far more
reasonable default, so we set that.

I would have assumed that ASSUME_CHARSET was the correct config value
here, based on the documentation [2], but that doesn't appear to work.

On the command-line, the equivalent flag appears to be
`--display_charset=utf-8`.

[1]: https://www.rfc-editor.org/rfc/rfc2616
[2]: https://lynx.invisible-island.net/lynx_help/body.html