tmux: add key binding to clear to scroll buffer
Adds a binding of <C-b>b to clear the scroll buffer, and the screen.

In macOS Terminal.app and in iTerm, Cmd-K clears the screen and the
scroll buffer. This is a pretty nice feature, particularly when
searching backwards through test results, since it avoids picking up
output from previous runs.