vim: remap <C-]> to <C-\\><C-n> in terminals
In neovim terminals, C-] (ESC) is passed through to the terminal, so the
only way to return to normal mode is <C-\><C-n>, which is particularly
difficult on a japanese keyboard, since there is no physical \ key, just
a ¥ key.

Instead, map C-] to behave as C-] would in non-terminal contexts.