gpg,git,mutt: use new gpg key
Finally got around to creating an ed25519 singing key to replace my
ancient RSA 2048 key. As evidence that the new key belongs to me, I've
signed it with my old key. For completeness, I've also signed the old
key with the new key.

This git commit should be signed with the old key but all further
commits will be signed with the new key.

This also adds a host-specific config for `hamachi`, and renames the
previous `yanagi` config to `ichou` (the host was renamed).